move the test files to _test_ folder
